refact: further decoupling components into files
